src/third_party/blink/web_tests/http/tests/js-self-profiling/foreign-script-cors-allow.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>

  <script crossorigin src="http://127.0.0.1:8080/js-self-profiling/resources/external-script.js"></script>
</head>
<body>
  <script>
    promise_test(async t => {
      assert_not_equals(window.externalScriptFunction, undefined);
      const trace = await ProfileUtils.profileFunction(externalScriptFunction);

      assert_true(ProfileUtils.containsFrame(trace, {
        name: 'externalScriptFunction',
        resourceId: trace.resources.indexOf('http://127.0.0.1:8080/js-self-profiling/resources/external-script.js'),
      }), 'foreign frame is present');

    }, 'function from foreign-origin script with CORS is included in trace');
  </script>
</body>
</html>
